server update + refactor

+++ server/server.c
@@ -6,21 +6,126 @@
 #include <sys/socket.h>
 #include <netinet/in.h>
 #include <string.h>
-#include <sys/epoll.h>
 #include <fcntl.h>
+#include <openssl/ssl.h>
+#include <openssl/err.h>
+#include <sys/epoll.h>
+#include "../threadsdata.h"
+#include "server.h"
+#include "jwt.h"
 #include "../utils/utils.h"
-
-#define PORT 7777
-#define INSOCK_QUEUE_SIZE 256
 
 // --------- server work logic in handlers
 
-void AskToken(const struct RequestParam_s *request, struct ResponseParam_s *response) {
+const char *jwtKey = "YW50aS1wYXR0ZXJu"; // anti-pattern
 
+
+
+void AskToken(const struct RequestParam_s *request, struct ResponseParam_s *response) {
+    char *user = LookUpParamValue(request->queryParams, "name");
+    if (user) {
+        printf("%s ask token\n", user);
+
+        // 1. search in DB user (we have a simple user data example without hash of user pass or smth else)
+
+
+
+        char *dbRes = "{\"_id\": \"277353dead\", \"username\": \"pupsik\", \"services\": [\"serv1\", \"serv2\"]}";
+
+        // 2. jwt
+
+        struct Jwt_s *jwt = CreateJwt(jwtKey, strlen(jwtKey));
+        AddJsonAsPayload(jwt, dbRes);
+
+        response->body = ConstructJwt(jwt);
+        if (response->body == NULL) {
+            response->status = INTERNAL_SERVER_ERROR;
+        } else {
+            SetContentType(response, "plain/text");
+            response->status = OK;
+        }
+
+        ReleaseJwt(jwt);
+    } else {
+        response->status = BAD_REQUEST;
+    }
 }
 
 void ValidateToken(const struct RequestParam_s *request, struct ResponseParam_s *response) {
+    if (request->body == NULL) {
+        response->status = BAD_REQUEST;
+        return;
+    }
+    char *body = "eyJhbGciOiAiSFMyNTYiLCAidHlwIjogIkpXVCJ9.eyJfaWQiOiAiMjc3MzUzZGVhZCIsICJ1c2VybmFtZSI6ICJwdXBzaWsiLCAic2VydmljZXMiOiBbInNlcnYxIiwgInNlcnYyIl19.0u12/m86MXmmiaT14q4w9YF8UM70AqFHOVSzq994uns=";
 
+    if (!CheckJwtHeader(body)) {
+        response->status = FORBIDDEN;
+        return;
+    }
+    if (!CheckJwtSign(body, jwtKey, strlen(jwtKey))) {
+        response->status = FORBIDDEN;
+        return;
+    }
+
+    // check jwt payload
+
+    // 1. get only payload
+    char *localBody = strdup(body);
+    strtok(localBody, "."); // skip header
+    char *payload = strtok(NULL, ".");
+    if (payload == NULL) {
+        free(localBody);
+        response->status = FORBIDDEN;
+        return;
+    }
+
+    // 2. base64 decode
+    int decPayloadLen = CalcDecodedLen(strlen(payload));
+    char *decodedPayload = malloc(decPayloadLen + 1);
+    memset(decodedPayload, 0, decPayloadLen + 1);
+    Base64Decode(payload, strlen(payload), decodedPayload);
+
+    // 3. check if it is json and this json has needed field
+
+    cJSON *payloadJson = cJSON_Parse(decodedPayload);
+    if (payloadJson == NULL) {
+        free(localBody);
+        free(decodedPayload);
+
+        response->status = FORBIDDEN;
+        return;
+    }
+
+    cJSON *userName = cJSON_GetObjectItem(payloadJson, "username");
+    if (!cJSON_IsString(userName) || userName->valuestring == NULL) {
+        cJSON_Delete(payloadJson);
+        free(decodedPayload);
+        free(localBody);
+
+        response->status = FORBIDDEN;
+        return;
+    }
+
+    // 4. search in DB user doc
+
+
+
+    char *dbRes = "{\"_id\": \"277353dead\", \"username\": \"pupsik\", \"services\": [\"serv1\", \"serv2\"]}";
+
+    // 5. compare DB result and token value
+    if (strcmp(decodedPayload, dbRes) != 0) {
+        cJSON_Delete(payloadJson);
+        free(decodedPayload);
+        free(localBody);
+
+        response->status = FORBIDDEN;
+        return;
+    }
+
+    cJSON_Delete(payloadJson);
+    free(decodedPayload);
+    free(localBody);
+    response->status = OK;
 }
 
 // --------- internal server functions
@@ -46,6 +151,79 @@ void RunMethod(struct ServerData_s *sd, const char *request, char *response) {
     ReleaseReq(rq);
 }
 
+int CreateSocket(int port) {
+    int listfd = 0;
+    struct sockaddr_in saddr;
+    saddr.sin_family = AF_INET;
+    saddr.sin_port = htons(port);
+    saddr.sin_addr.s_addr = htons(INADDR_ANY);
+
+    listfd = socket(AF_INET, SOCK_STREAM, 0);
+    if(listfd == -1) {
+        printf("socket() ret %d, %s\nstop child thread\n", errno, strerror(errno));
+        return -1;
+    }
+
+    // set non-blocking
+    int flags;
+    if ((flags = fcntl(listfd, F_GETFL, 0)) < 0) {
+        close(listfd);
+
+        printf("fcntl(F_GETFL) ret %d, %s\n", errno, strerror(errno));
+        return -1;
+    }
+    if (fcntl(listfd, F_SETFL, flags | O_NONBLOCK) < 0) {
+        close(listfd);
+
+        printf("fcntl(F_SETFL) ret %d, %s\n", errno, strerror(errno));
+        return -1;
+    }
+
+    if(bind(listfd, (const struct sockaddr *) &saddr, sizeof(saddr)) == -1) {
+        close(listfd);
+
+        printf("bind() ret %d, %s\nstop child thread\n", errno, strerror(errno));
+        return -1;
+    }
+
+    if (listen(listfd, INSOCK_QUEUE_SIZE) == -1) {
+        close(listfd);
+
+        printf("listen() ret %d %s\nstop child thread\n", errno, strerror(errno));
+        return -1;
+    }
+
+    return listfd;
+}
+
+SSL_CTX *CreateSslContext() {
+    const SSL_METHOD *method;
+    SSL_CTX *ctx;
+
+    method = TLS_server_method();
+    ctx = SSL_CTX_new(method);
+    if (!ctx) {
+        printf("Unable to create SSL context\n");
+        return NULL;
+    }
+    return ctx;
+}
+
+int ConfigureContext(SSL_CTX *ctx) {
+    /* set key and cert */
+    if (SSL_CTX_use_certificate_file(ctx, "../server_keys/server-cert.pem", SSL_FILETYPE_PEM) <= 0) {
+        ERR_print_errors_fp(stderr);
+        return 0;
+    }
+    if (SSL_CTX_use_PrivateKey_file(ctx, "../server_keys/server-key.pem", SSL_FILETYPE_PEM) <= 0) {
+        ERR_print_errors_fp(stderr);
+        return 0;
+    }
+    SSL_CTX_set_min_proto_version(ctx, TLS1_2_VERSION); // no SSLv2, SSL_v3, TLSv1.0, TLSv1.1
+
+    return 1;
+}
+
 void *NetInterface (void *arg) {
     struct ThreadData_s *threadData = (struct ThreadData_s*)arg;
 
@@ -60,48 +238,21 @@ void *NetInterface (void *arg) {
 
     int listfd = 0;
     int connfd = 0;
-    struct sockaddr_in saddr;
+    SSL_CTX *ctx;
 
-    saddr.sin_family = AF_INET;
-    saddr.sin_port = htons(PORT);
-    saddr.sin_addr.s_addr = htons(INADDR_ANY);
-
-    listfd = socket(AF_INET, SOCK_STREAM, 0);
-    if(listfd == -1) {
-        printf("socket() ret %d, %s\nstop child thread\n", errno, strerror(errno));
+    ctx = CreateSslContext();
+    if (ctx == NULL) {
         kill(threadData->mainPid, SIGUSR1);
         pthread_exit(threadData);
     }
 
-    // set non-blocking
-    int flags;
-    if ((flags = fcntl(listfd, F_GETFL, 0)) < 0) {
-        close(listfd);
-
-        printf("fcntl(F_GETFL) ret %d, %s\n", errno, strerror(errno));
-        kill(threadData->mainPid, SIGUSR1);
-        pthread_exit(threadData);
-    }
-    if (fcntl(listfd, F_SETFL, flags | O_NONBLOCK) < 0) {
-        close(listfd);
-
-        printf("fcntl(F_SETFL) ret %d, %s\n", errno, strerror(errno));
+    if (!ConfigureContext(ctx)) {
         kill(threadData->mainPid, SIGUSR1);
         pthread_exit(threadData);
     }
 
-    if(bind(listfd, (const struct sockaddr *) &saddr, sizeof(saddr)) == -1) {
-        close(listfd);
-
-        printf("bind() ret %d, %s\nstop child thread\n", errno, strerror(errno));
-        kill(threadData->mainPid, SIGUSR1);
-        pthread_exit(threadData);
-    }
-
-    if (listen(listfd, INSOCK_QUEUE_SIZE) == -1) {
-        close(listfd);
-
-        printf("listen() ret %d %s\nstop child thread\n", errno, strerror(errno));
+    listfd = CreateSocket(PORT);
+    if (listfd == -1) {
         kill(threadData->mainPid, SIGUSR1);
         pthread_exit(threadData);
     }
@@ -151,26 +302,44 @@ void *NetInterface (void *arg) {
                     continue;
                 }
 
-                ssize_t n = read(connfd, recBuf, sizeof(recBuf));
-                if(n == -1) {
-                    close(connfd);
-                    printf("read() ret %d, %s\n", errno, strerror(errno));
-                    continue;
-                }
+                SSL *sslConnection;
+                sslConnection = SSL_new(ctx);
+                SSL_set_fd(sslConnection, connfd);
 
-                snprintf(sendBuf, sizeof(sendBuf), "server got msg: %s\n", recBuf);
-                printf("%s", sendBuf);
+                if (SSL_accept(sslConnection) <= 0) {
+                    ERR_print_errors_fp(stderr);
+                } else {
 
-                RunMethod(&threadData->sd, recBuf, sendBuf);
+                    //ssize_t n = read(connfd, recBuf, sizeof(recBuf));
+                    ssize_t n = SSL_read(sslConnection, recBuf, sizeof(recBuf));
+                    if(n == -1) {
+                        SSL_shutdown(sslConnection);
+                        SSL_free(sslConnection);
+                        close(connfd);
+                        ERR_print_errors_fp(stderr);
+                        continue;
+                    }
 
-                n = write(connfd, sendBuf, strlen(sendBuf));
-                if (n == -1) {
-                    close(connfd);
-                    printf("write() ret %d, %s\n", errno, strerror(errno));
-                    kill(threadData->mainPid, SIGUSR1);
-                    break;
-                }
+                    snprintf(sendBuf, sizeof(sendBuf), "server got msg: %s\n", recBuf);
+                    printf("%s", sendBuf);
 
+                    RunMethod(&threadData->sd, recBuf, sendBuf);
+
+                    //n = write(connfd, sendBuf, strlen(sendBuf));
+                    n = SSL_write(sslConnection, sendBuf, strlen(sendBuf));
+                    if (n == -1) {
+                        SSL_shutdown(sslConnection);
+                        SSL_free(sslConnection);
+                        close(connfd);
+                        ERR_print_errors_fp(stderr);
+
+                        kill(threadData->mainPid, SIGUSR1);
+                        break;
+                    }
+                } // if ssl_accept
+
+                SSL_shutdown(sslConnection);
+                SSL_free(sslConnection);
                 close(connfd);
 
             } // if our connection
@@ -180,6 +349,7 @@ void *NetInterface (void *arg) {
     printf("stop child thread\n");
 
     close(listfd);
+    SSL_CTX_free(ctx);
     FreeHandlersList(threadData->sd.handlres);
     pthread_exit(threadData);
 }
